requirements.txt
streamlit
gTTS
pyttsx3
pytesseract
tesseract
google.generativeai
langchain_google_genai
Pillow